video_checker/requirements.txt
torch==2.8.0
torchvision==0.15.1   # compatible with torch 2.8.0
opencv-python-headless==4.9.0.74
numpy==1.26.0
python-magic==0.4.27
tensorflow==2.14.0
keras==2.14.0
